lint fix: disable abc for library
Unfortunately we have to disable it at this point, if we don't want to risk writing more functions to cover the same space. Even if the buildup structure of the hashmap is completely removed from these functions (thus reducing the total number of assignments considerably), we are left with a slightly over AbcSize for get_macs, which is due to the OS-checking inside. So ignore it unless we find a better solution

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>